+++ R/ggpacket_show.R
@@ -131,12 +131,11 @@ format_ggpacket_mapping.NULL <- function(x,
 }
 
 #' @importFrom utils capture.output
-#' @importFrom crayon red
 format_ggpacket_mapping.default <- function(x,
     width = getOption("width", 80) * 0.9, missing_aes = character(0L)) {
 
   x[missing_aes] <- " MISSING "
-  gsub("\" MISSING \"", crayon::red("<missing>"), utils::capture.output(x)[-1])
+  gsub("\" MISSING \"", "<missing>", utils::capture.output(x)[-1])
 }
 
 format_ggpacket_ggcalls <- function(x,